+++ src/components/Process/index.js
@@ -1,0 +1,5 @@
+export { default } from './ProcessInfo'
+export { default as ProcessInfo } from './ProcessInfo'
+export { default as ProcessInfoRow } from './ProcessInfoRow'
+export { default as ProcessStatusSummary } from './ProcessStatusSummary'
+export { default as ProcessStatusInfo } from './ProcessStatusInfo'